+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 30.0-SNAPSHOT - unreleased
 
+### 🎁 New Features
+
+* Support for reporting App Build, TabID and Client ID in websocket admin page
+
 ## 29.1.0 - 2025-04-08
 
 ### 🎁 New Features